Parse grok files and extract their patterns.
This is the first step - being able to specify a pattern by name and get
the result back. This is not complete or usable yet though, because it
doesn't perform the subpattern substitution that grok does, and it only
outputs in oniguruma compatible regex.

The next few goals are to substitute the %{pattern interpolation}
syntaxes so that plain regular expressions are returned, and then add a
way to convert the existing regex to other flavor of regex, such as
Python.